# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Executable Governance API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations:
    get:
      operationId: listOrganizations
      tags: [organizations]
      summary: List all organizations
      responses:
        "200":
          description: List of organizations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Organization"
    post:
      operationId: createOrganization
      tags: [organizations]
      summary: Create an organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationBody"
      responses:
        "201":
          description: Created organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
  /organizations/{id}:
    get:
      operationId: getOrganization
      tags: [organizations]
      summary: Get an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateOrganization
      tags: [organizations]
      summary: Update an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationBody"
      responses:
        "200":
          description: Updated organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteOrganization
      tags: [organizations]
      summary: Delete an organization
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Policies
...
  /analytics/summary:
    get:
      operationId: getAnalyticsSummary
      tags: [analytics]
      summary: Get platform summary stats
      parameters:
        - name: organizationId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Summary statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummary"
  /analytics/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [analytics]
      summary: Get recent rule and policy activity
      parameters:
        - name: organizationId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Recent activity feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
  /analytics/policy-breakdown:
    get:
      operationId: getPolicyBreakdown
      tags: [analytics]
      summary: Get policy status breakdown by organization
      parameters:
        - name: organizationId
          in: query
          required: false
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Policy breakdown
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PolicyBreakdown"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # Organization schemas
    Organization:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - createdAt
        - updatedAt
    CreateOrganizationBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
      required:
        - name
    UpdateOrganizationBody:
      type: object
      properties:
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]

    # Policy schemas
...
    AnalyticsSummary:
      type: object
      properties:
        totalOrganizations:
          type: integer
        totalPolicies:
          type: integer
        publishedPolicies:
          type: integer
        draftPolicies:
          type: integer
        totalRules:
          type: integer
        publishedRules:
          type: integer
        draftRules:
          type: integer
        totalUsers:
          type: integer
      required:
        - totalOrganizations
        - totalPolicies
        - publishedPolicies
        - draftPolicies
        - totalRules
        - publishedRules
        - draftRules
        - totalUsers
    ActivityItem:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [rule_created, rule_updated, rule_published, policy_created, policy_updated, policy_published, policy_archived]
        entityId:
          type: integer
        entityName:
          type: string
        entityType:
          type: string
          enum: [rule, policy]
        organizationName:
          type: string
        policyName:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - entityId
        - entityName
        - entityType
        - organizationName
        - createdAt
    PolicyBreakdown:
      type: object
      properties:
        totalPolicies:
          type: integer
        byStatus:
          type: object
          properties:
            draft:
              type: integer
            published:
              type: integer
            archived:
              type: integer
          required:
            - draft
            - published
            - archived
        byDomain:
          type: array
          items:
            type: object
            properties:
              domain:
                type: string
              count:
                type: integer
              publishedCount:
                type: integer
            required:
              - domain
              - count
              - publishedCount
      required:
        - totalPolicies
        - byStatus
        - byDomain